Initial commit
The README and COPYING files are included to start the repo.

The `todo` file is the frontend of the program, which is in charge of
understanding which action you want to do (e.g. add, list, etc).

The `todo_modules` directory contains all the scripts that make up the
program. `todo` automatically defers tasks to these scripts.

The modules include:

- `add.sh`: Add a todo item
- `done.sh`: Mark a todo item as done
- `help.sh`: Print usage text
- `list.sh`: List and filter todo items
- `tags.sh`: List tags such as contexts or projects